fix(infraclientx): move from sessions to authn directory
