+++ authn/webapi/sessions/hooks/mutations/mutations.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 	"net/http"
 
-	"webapi/sessions/infraclientx/fetchx"
-	fetchRequests "webapi/sessions/infraclientx/fetchx/requests"
+	"webapi/infraclientx/fetchx"
+	fetchRequests "webapi/infraclientx/fetchx/requests"
 	"webapi/sessions/hooks/errors"
 	"webapi/sessions/hooks/requests"
 	"webapi/sessions/hooks/responses"
