+++ authn/webapi/store/users/hooks/mutations/mutations.go
@@ -10,7 +10,7 @@ import (
 	"webapi/store/users/hooks/requests"
 	"webapi/store/users/hooks/responses"
 
-	"webapi/sessions/infraclientx/verifyx"
+	"webapi/infraclientx/verifyx"
 )
 
 func writeUsersResponse(w http.ResponseWriter, users *controller.SafeUsers) {
