+++ authn/webapi/store/users/hooks/queries/queries.go
@@ -10,7 +10,7 @@ import (
 	"webapi/store/users/hooks/requests"
 	"webapi/store/users/hooks/responses"
 
-	"webapi/sessions/infraclientx/verifyx"
+	"webapi/infraclientx/verifyx"
 )
 
 func writeUsersResponse(w http.ResponseWriter, users *controller.SafeUsers) {
